+++ app.config.js
@@ -1,5 +1,5 @@
 // 引入自定义公共函数
-import myPubFunction from '@/common/function/myPubFunction.js'
+import myPubFunction from '@/common/js/myPubFunction.js'
 export default {
   // 开发模式启用调式模式(请求时会打印日志)
   debug: process.env.NODE_ENV !== 'production',
