+++ changelog.md
@@ -83,7 +83,7 @@
 - 使用`uni.getUniverifyManager`优化一键登陆中，点击第三方登陆的逻辑：未勾选隐私政策时，toast提醒并阻止了一键登陆界面的close
 - 新增支持看激励视频广告签到
 ## 1.1.22（2021-11-10）
-删除`common/openApp.js`中可选链操作符，解决vue3版本在hbuilderX内置浏览器不兼容的问题
+删除`common/js/openApp.js`中可选链操作符，解决vue3版本在hbuilderX内置浏览器不兼容的问题
 ## 1.1.21（2021-11-10）
 新增app端列表页面使用原生list下拉刷新
 ## 1.1.20（2021-11-08）
@@ -155,7 +155,7 @@
 默认首页为nvue页面+fast
 ## 1.0.44（2021-08-05）
 解决首页为非nvue页面时白屏的问题。
-- 注意：本次在`common/appInit.js`中修改了路由拦截的逻辑，是个兼容方案；当首页为非nvue页面，路由拦截器逻辑会在加载首页时执行。接下来新版本的hx编译的uni-app项目无论首页是否为nvue都不走拦截器，保持各端逻辑一致。
+- 注意：本次在`common/js/appInit.js`中修改了路由拦截的逻辑，是个兼容方案；当首页为非nvue页面，路由拦截器逻辑会在加载首页时执行。接下来新版本的hx编译的uni-app项目无论首页是否为nvue都不走拦截器，保持各端逻辑一致。
 ## 1.0.43（2021-08-02）
 1. 微信小程序端，新增：微信登录成功后，弹出是否"获取微信头像和昵称，完善个人资料"的弹框
 2. APP端，新增逻辑：微信登录成功后，自动获取用户的微信昵称和头像完善用户个人资料
